refactor: reverts footer to use flex
